Rename tex.BufferX new function.
1. new function names are more comprehensible.
